+++ nicegui/static/nicegui.css
@@ -68,7 +68,7 @@
   width: 100%;
   height: 16rem;
 }
-.nicegui-jsoneditor {
+.nicegui-json-editor {
   width: 100%;
   height: 16rem;
 }
